% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n1qn1.R
\name{.n1qn1ptr}
\alias{.n1qn1ptr}
\title{This gives the function pointers in the n1qn1 library}
\usage{
.n1qn1ptr()
}
\value{
list of pointers to the n1qn1 functions
}
\description{
Using this will allow C-level linking by function pointers instead
of abi.
}
\examples{
.n1qn1ptr()
}
\author{
Matthew L. Fidler
}
